delete user msg command start from !
