+++ pkg/ceviord/const.go
@@ -1,0 +1,5 @@
+package ceviord
+
+const prefix = "!"
+const strLenMax = 300
+const DiscordPostLenLimit = 2000
